+++ packages/column-views/stories/correlation-chart/correlation-chart.stories.ts
@@ -1,12 +1,13 @@
 import { Meta } from "@storybook/react-vite";
 import "@macrostrat/style-system";
-import { useCorrelationLine } from "./utils";
+import { useCorrelationLine, CorrelationColumnHeader } from "./utils.ts";
 import {
   ColumnCorrelationMap,
   ColumnCorrelationProvider,
   MergeSectionsMode,
   useCorrelationMapStore,
-} from "../..";
+  useColumnMapLink,
+} from "../../src";
 import { hyperStyled } from "@macrostrat/hyper";
 import {
   MacrostratDataProvider,
@@ -15,10 +16,13 @@ import {
 } from "@macrostrat/data-provider";
 
 import styles from "./stories.module.sass";
-import { CorrelationChart, CorrelationChartProps } from "../main";
+import {
+  CorrelationChart,
+  CorrelationChartProps,
+} from "../../src/correlation-chart/main.ts";
 import { ErrorBoundary, useAsyncMemo } from "@macrostrat/ui-components";
 import { OverlaysProvider } from "@blueprintjs/core";
-import { EnvironmentColoredUnitComponent } from "../../units";
+import { EnvironmentColoredUnitComponent } from "../../src/units";
 import { scaleLinear, scalePow } from "d3-scale";
 import { MacrostratInteractionProvider } from "@macrostrat/data-components";
 
@@ -78,6 +82,8 @@ function CorrelationDiagramWrapper(props: Omit<CorrelationChartProps, "data">) {
   const focusedColumns = useCorrelationMapStore(
     (state) => state.focusedColumns,
   );
+  // Link column hover/click to the map (highlight on hover, frame on click)
+  const columnMapLink = useColumnMapLink();
 
   const columnUnits = useAsyncMemo(async () => {
     const col_ids = focusedColumns.map((col) => col.properties.col_id);
@@ -88,7 +94,11 @@ function CorrelationDiagramWrapper(props: Omit<CorrelationChartProps, "data">) {
     h(
       ErrorBoundary,
       h(OverlaysProvider, [
-        h(CorrelationChart, { data: columnUnits, ...props }),
+        h(CorrelationChart, {
+          data: columnUnits,
+          ...props,
+          ...columnMapLink,
+        }),
       ]),
     ),
   ]);
@@ -161,6 +171,11 @@ export default {
         type: "number",
       },
     },
+    minPixelScale: {
+      control: {
+        type: "number",
+      },
+    },
     targetUnitHeight: {
       control: {
         type: "number",
@@ -180,6 +195,10 @@ export default {
       control: {
         type: "number",
       },
+    },
+    columnHeaderComponent: {
+      control: false,
+      table: { category: "Content" },
     },
   },
 } as Meta<typeof CorrelationStoryUI>;
@@ -226,3 +245,35 @@ eODPCorrelationChart.args = {
   focusedLine: "-125,38 -120,32",
   projectID: 3,
 };
+
+export const WithColumnHeaders = Template.bind({});
+WithColumnHeaders.args = {
+  columnHeaderComponent: CorrelationColumnHeader,
+};
+WithColumnHeaders.parameters = {
+  docs: {
+    description: {
+      story:
+        "Arbitrary content (here, the column name and ID) can be placed above " +
+        "each column using the `columnHeaderComponent` prop. Headers stay " +
+        "pinned to the top of the chart while scrolling.",
+    },
+  },
+};
+
+export const WideColumnSpacing = Template.bind({});
+WideColumnSpacing.args = {
+  columnSpacing: 30,
+  columnWidth: 90,
+  columnHeaderComponent: CorrelationColumnHeader,
+};
+WideColumnSpacing.parameters = {
+  docs: {
+    description: {
+      story:
+        "The `columnSpacing` prop controls the horizontal gap between columns. " +
+        "Column headers remain aligned with their columns regardless of spacing.",
+    },
+  },
+};
+