+++ packages/timescale/src/index.ts
@@ -20,7 +20,11 @@ import { ReactNode, useCallback, useMemo, useRef } from "react";
 import h from "./main.module.sass";
 
 export * from "./intervals-api";
-export type { Interval } from "./types";
+export type {
+  Interval,
+  TimescaleClickHandler,
+  TimescaleClickData,
+} from "./types";
 export {
   IncreaseDirection,
   TimescaleOrientation,